// src/enums/error-keys.ts
export enum ErrorKeys {
  MissingGeoServiceProvider = "MissingGeoServiceProvider",
  InvalidGeoServiceProvider = "InvalidGeoServiceProvider",
  MissingGoogleApiKey = "MissingGoogleApiKey",
  MissingSmartyCredentials = "MissingSmartyCredentials",
  InternalServerError = "InternalServerError",
  BadRequestError = "BadRequestError",
  PayloadValidationError = "PayloadValidationError",
  RateLimitExceeded = "RateLimitExceeded",
  SmartyStreetsApiError = "SmartyStreetsApiError",
  GoogleMapsApiError = "GoogleMapsApiError",
  MissingGoogleMapsURL = "MissingGoogleMapsURL",
  MissingSmartyURL = "MissingSmartyURL",
}